+++ panoptes/utils/tests/test_utils.py
@@ -10,28 +10,6 @@ from panoptes.utils.library import load_c_library
 from panoptes.utils.logger import get_root_logger
 
 
-def test_error(capsys):
-    with pytest.raises(error.PanError) as e_info:
-        raise error.PanError(msg='Testing message')
-
-    assert str(e_info.value) == 'PanError: Testing message'
-
-    with pytest.raises(error.PanError) as e_info:
-        raise error.PanError()
-
-    assert str(e_info.value) == 'PanError'
-
-    with pytest.raises(SystemExit) as e_info:
-        raise error.PanError(msg="Testing exit", exit=True)
-    assert e_info.type == SystemExit
-    assert capsys.readouterr().out.strip() == 'TERMINATING: Testing exit'
-
-    with pytest.raises(SystemExit) as e_info:
-        raise error.PanError(exit=True)
-    assert e_info.type == SystemExit
-    assert capsys.readouterr().out.strip() == 'TERMINATING: No reason specified'
-
-
 def test_bad_load_module():
     with pytest.raises(error.NotFound):
         load_module('FOOBAR')
